# services/stash-service/api-spec/src/main/resources/openapi/stash-api.yaml
openapi: 3.0.3
info:
  title: Linkpouch Stash Service API
  description: API for managing stashes and links in Linkpouch
  version: 1.0.0
  contact:
    name: Linkpouch Team
servers:
  - url: /api
    description: Local development
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived JWT issued by POST /stashes/{stashId}/access-token
  schemas:
...
    AiSettingsResponse:
      type: object
      properties:
        provider:
          type: string
          enum:
            - NONE
            - OPENROUTER_INCLUDED
            - OPENROUTER
            - OPENAI
            - ANTHROPIC
            - OPENCODE
        model:
          type: string
          nullable: true
        hasApiKey:
          type: boolean
          description: True when an encrypted API key is stored (never returns the raw key)
        customPrompt:
          type: string
          nullable: true
          description: Custom system prompt to use instead of the default
    UpsertAiSettingsRequest:
      type: object
      required:
        - provider
      properties:
        provider:
          type: string
          enum:
            - NONE
            - OPENROUTER_INCLUDED
            - OPENROUTER
            - OPENAI
            - ANTHROPIC
            - OPENCODE
        apiKey:
          type: string
          nullable: true
          description: API key for the provider (null preserves an existing key; omit for OPENROUTER_INCLUDED)
        model:
          type: string
          nullable: true
        customPrompt:
          type: string
          nullable: true
          description: Custom system prompt to use instead of the default
...
    AiModelInfo:
      type: object
      required: [id]
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        contextLength:
          type: integer
          nullable: true
          description: Maximum context window in tokens
        pricingPrompt:
          type: number
          format: double
          nullable: true
          description: Cost per input token in USD (multiply × 1,000,000 for display)
        pricingCompletion:
          type: number
          format: double
          nullable: true
          description: Cost per output token in USD
    AiModelsResponse:
      type: object
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/AiModelInfo'
          description: Available models for the given provider
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        errorCode:
          type: string
          description: Machine-readable error code (e.g. PASSWORD_REQUIRED, INVALID_TOKEN)
        path:
          type: string
        signatureRefreshedAt:
          type: string
          format: date-time
          description: When the stash signature was last regenerated (only set when errorCode is SIGNATURE_REGENERATED)
...
    AiCredentialsResponse:
      type: object
      required:
        - apiKey
      properties:
        apiKey:
          type: string
          description: Decrypted AI API key for the account that owns this stash, or empty string if none configured
paths:
...
  /account/ai-settings:
    get:
      summary: Get AI settings for the authenticated account
      operationId: getAiSettings
      tags:
        - AccountAiSettings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: AI settings for the account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiSettingsResponse'
        '404':
          description: No AI settings have been configured yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or missing account token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Create or update AI settings (use provider NONE to disable)
      operationId: upsertAiSettings
      tags:
        - AccountAiSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertAiSettingsRequest'
      responses:
        '200':
          description: AI settings saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiSettingsResponse'
        '401':
          description: Invalid or missing account token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /account/ai-settings/models:
    get:
      summary: Fetch available model IDs for a given provider and API key (stateless, not persisted)
      operationId: getAiModels
      tags:
        - AccountAiSettings
      security:
        - bearerAuth: []
      parameters:
        - name: provider
          in: query
          required: true
          schema:
            type: string
            enum:
              - OPENROUTER_INCLUDED
              - OPENROUTER
              - OPENAI
              - ANTHROPIC
              - OPENCODE
        - name: X-Ai-Api-Key
          in: header
          required: false
          schema:
            type: string
          description: API key to use for fetching models (ignored for OPENROUTER_INCLUDED/OPENROUTER public list). Sent as a header to avoid leaking into server logs and browser history.
      responses:
        '200':
          description: List of available model IDs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiModelsResponse'
        '400':
          description: Invalid provider or API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or missing account token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stashes/{stashId}/ai-credentials:
    parameters:
      - name: stashId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Internal — fetch decrypted AI API key for a stash's account (indexer only)
      description: >
        Called by the indexer at processing time to retrieve the AI API key for the account
        that owns the given stash. Requires the shared X-Indexer-Secret header.
        This endpoint is blocked at the API gateway and only reachable from within the cluster.
      operationId: getAiCredentials
      tags:
        - AccountAiSettings
      parameters:
        - name: X-Indexer-Secret
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: AI credentials for the account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiCredentialsResponse'
        '401':
          description: Invalid indexer secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...